Add hook that fetches data of parking charge zones
